requirements.txt
Django==3.2.16
django-mptt==0.14.0
